Cast size_t to unsigned int for printf()
git-svn-id: https://svn.blender.org/svnroot/verse/trunk@2578 49c17534-1d02-dc11-b283-0030488c597c

+++ purple/plugins.c
@@ -592,7 +592,7 @@ PluginStatus plugin_instance_compute(PInstance *inst)
 				if((o = inst->resolver(module, inst->resolver_data)) != NULL)
 					ps->port[i] = o;
 				else
-					printf("Couldn't resolve port %u\n", i);
+					printf("Couldn't resolve port %u\n", (unsigned int) i);
 			}
 			else
 				ps->port[i] = &ps->input[i];
